+++ src/ApiManager/__tests__/@ApiManager.core.test.ts
@@ -18,7 +18,7 @@ import {
 import type { TJsSIP } from '@/types';
 
 // Мокаем logger
-jest.mock('../../logger', () => {
+jest.mock('@/logger', () => {
   return jest.fn();
 });