+++ src/ApiManager/constants.ts
@@ -38,7 +38,6 @@ export enum EShareState {
 export enum EParticipantType {
   SPECTATOR = 'SPECTATOR',
   PARTICIPANT = 'PARTICIPANT',
-  SPECTATOR_OVER_SFU = 'SPECTATOROVERSFU',
 }
 
 export enum EContentTypeReceived {