+++ src/CallManager/index.ts
@@ -1,7 +1,8 @@
 export { default as CallManager } from './@CallManager';
 export { ECallCause } from './causes';
-export { EEvent as ECallEvent, Originator } from './eventNames';
+export type { TEventMap } from './eventNames';
+export { EEvent as ECallEvent } from './eventNames';
 export type { TEvent as TCallEvent, TEvents as TCallEvents } from './eventNames';
 export { default as hasCanceledCallError } from './hasCanceledCallError';
 export { MCUCallStrategy } from './MCUCallStrategy';
-export type { TCustomError, TGetServerUrl, ITransceiverStorage } from './types';
+export type { TCustomError, TGetServerUrl } from './types';